--- .pre-commit-config.yaml
@@ -1,7 +1,0 @@
-fail_fast: false
-repos:
--   repo: https://github.com/pre-commit/mirrors-clang-format
-    rev: 'v16.0.3'
-    hooks:
-      - id: clang-format
-        exclude_types: [javascript, json]
