+++ cmake/gwtoolboxdll_plugins.cmake
@@ -1,11 +1,5 @@
 macro(add_tb_plugin PLUGIN)
-  if(CMAKE_BUILD_TYPE MATCHES Debug)
-    set(PLUGIN_NAME ${PLUGIN}_debug)
-  else()
-    set(PLUGIN_NAME ${PLUGIN})
-  endif()
-
-  add_library(${PLUGIN_NAME} SHARED)
+  add_library(${PLUGIN} SHARED)
 
   file(
     GLOB
@@ -18,10 +12,11 @@ macro(add_tb_plugin PLUGIN)
     ${gwtoolboxpp_SOURCE_DIR}/plugins/Base/*.cpp)
 
   target_include_directories(
-    ${PLUGIN_NAME} PRIVATE ${gwtoolboxpp_SOURCE_DIR}/plugins/Base
-                           ${gwtoolboxpp_SOURCE_DIR}/GWToolboxdll)
+    ${PLUGIN} PRIVATE ${gwtoolboxpp_SOURCE_DIR}/plugins/Base
+                      ${gwtoolboxpp_SOURCE_DIR}/GWToolboxdll
+                      src/${PLUGIN})
   target_link_libraries(
-    ${PLUGIN_NAME}
+    ${PLUGIN}
     PUBLIC RestClient
            directxtex
            easywsclient
@@ -30,20 +25,19 @@ macro(add_tb_plugin PLUGIN)
            imgui::fonts
            helper_utils)
 
-  target_compile_options(${PLUGIN_NAME} PRIVATE /W4 /Gy)
-  target_compile_options(${PLUGIN_NAME} PRIVATE $<$<NOT:$<CONFIG:Debug>>:/GL>)
-  target_compile_options(${PLUGIN_NAME} PRIVATE $<$<CONFIG:Debug>:/ZI /Od
-                                                /RTCs>)
-  target_compile_options(${PLUGIN_NAME} PRIVATE $<$<CONFIG:RelWithDebInfo>:/Zi>)
+  target_compile_options(${PLUGIN} PRIVATE /W4 /Gy)
+  target_compile_options(${PLUGIN} PRIVATE $<$<NOT:$<CONFIG:Debug>>:/GL>)
+  target_compile_options(${PLUGIN} PRIVATE $<$<CONFIG:Debug>:/ZI /Od /RTCs>)
+  target_compile_options(${PLUGIN} PRIVATE $<$<CONFIG:RelWithDebInfo>:/Zi>)
 
   set_target_properties(
-    ${PLUGIN_NAME}
+    ${PLUGIN}
     PROPERTIES VS_GLOBAL_RunCodeAnalysis false
                VS_GLOBAL_EnableMicrosoftCodeAnalysis true
                VS_GLOBAL_EnableClangTidyCodeAnalysis false)
 
   target_link_options(
-    ${PLUGIN_NAME}
+    ${PLUGIN}
     PRIVATE
     /WX
     /OPT:REF
@@ -59,15 +53,15 @@ macro(add_tb_plugin PLUGIN)
   )
 
   source_group(TREE "${CMAKE_CURRENT_SOURCE_DIR}" FILES ${SOURCES})
-  target_sources(${PLUGIN_NAME} PRIVATE ${SOURCES})
+  target_sources(${PLUGIN} PRIVATE ${SOURCES})
   target_compile_definitions(
-    ${PLUGIN_NAME} PRIVATE "NOMINMAX" "_WIN32_WINNT=_WIN32_WINNT_WIN7"
-                           "WIN32_LEAN_AND_MEAN" "VC_EXTRALEAN" "BUILD_DLL")
+    ${PLUGIN} PRIVATE "NOMINMAX" "_WIN32_WINNT=_WIN32_WINNT_WIN7"
+                      "WIN32_LEAN_AND_MEAN" "VC_EXTRALEAN" "BUILD_DLL")
 
-  set_target_properties(${PLUGIN_NAME} PROPERTIES FOLDER "src")
+  set_target_properties(${PLUGIN} PROPERTIES FOLDER "src")
 
   if(${ENABLE_WARNINGS})
-    target_set_warnings(TARGET ${PLUGIN_NAME} ENABLE ${ENABLE_WARNINGS}
-                        AS_ERRORS ${ENABLE_WARNINGS_AS_ERRORS})
+    target_set_warnings(TARGET ${PLUGIN} ENABLE ${ENABLE_WARNINGS} AS_ERRORS
+                        ${ENABLE_WARNINGS_AS_ERRORS})
   endif()
 endmacro()
