+++ src/_Lib/Helper/HeroSmartSkills.h
@@ -2,23 +2,7 @@ namespace HeroSmartSkills
 {
 void AttackTarget();
 
-bool UseBipOnPlayer();
-
-bool UseSplinterOnPlayer();
-
-bool UseVigSpiritOnPlayer();
-
-bool UseHonorOnPlayer();
-
-bool UseShelterInFight();
-
-bool UseUnionInFight();
-
-bool UseSosInFight();
-
 bool ShatterImportantHexes();
-
-bool RemoveImportantConditions();
 
 bool RuptEnemies();
